usewindowsize hook added inside cv downloader
