+++ app/Components/Navbar/CvDownloader.tsx
@@ -1,8 +1,31 @@
 'use client'
 import { cvDownloadHandler } from "@/app/lib/utils"
 import { useRouter } from 'next/navigation'
-import {useWindowSize} from '../../lib/useWindowSize'
 
+import { useState, useEffect } from 'react';
+
+export const useWindowSize = () => {
+  const [windowSize, setWindowSize] = useState({
+    width: 0,
+    height: 0,
+  });
+
+  useEffect(() => {
+    const handleResize = () => {
+      setWindowSize({
+        width: window.innerWidth,
+        height: window.innerHeight,
+      });
+    };
+
+    window.addEventListener('resize', handleResize);
+    handleResize();
+
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return windowSize;
+};
 
 export default function CvDownloader() {
   const { width } = useWindowSize();
